% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.installed.R
\name{is.installed}
\alias{is.installed}
\title{Is a package installed}
\usage{
is.installed(mypkg)
}
\arguments{
\item{mypkg}{Character vector. Package to be checked.}
}
\value{
Boolean. Is TRUE if the package is installed and available.
}
\description{
\code{is.installed} returns a Boolean describing if a package is installed.
}
\keyword{internal}
