% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nddspiro.quality.R
\name{nddspiro.quality}
\alias{nddspiro.quality}
\title{Get spirometry quality}
\usage{
nddspiro.quality(fev, fvc)
}
\arguments{
\item{fev}{Vector of double. One or more values of FEV1.}

\item{fvc}{Vector of double. One or more values of FVC.}
}
\value{
Character representing quality (possibilities are A,B,C,D,F).
}
\description{
Gives quality measure from series of spirometry measures
}
\keyword{internal}
