% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nddspiro.read.R
\name{nddspiro.read}
\alias{nddspiro.read}
\title{Import and process spirometry data}
\usage{
nddspiro.read(xmlfilename, indices = c("FEV1", "FVC", "FEV1_FVC", "PEFR"),
  traces = 2, accepted_only = TRUE, add_GLI = FALSE)
}
\arguments{
\item{xmlfilename}{Character vector. Filename of the XML file to process.}

\item{indices}{Optional. List of character vectors. The indices to be returned. Default is FEV1, FVC, FEV1FVC, PEFR. Possible values are "FEV1", "FEV6", "FVC", "FEV1_FVC", "PEFR", "FEF2575", "FET", "BEV", "EOTV", "VCmax"}

\item{traces}{Integer. Optional. Results per test to be returned (in descending order from Rank 1). Default=2.}

\item{accepted_only}{Optional. Boolean. If TRUE only "accepted" traces will be returned. Default=TRUE.}

\item{add_GLI}{Optional. Boolean. If TRUE will add GLI 2012 reference ranges. Requires rspiro package. Default=FALSE.}
}
\value{
Dataframe of test results
}
\description{
Extracts spirometry data from ndd Easyware XML files
}
\details{
This is the only function you are likely to need.
Can be used with\pkg{rspiro} from devtools::install_git(\"https://github.com/thlytras/rspiro.git\")
}
\examples{
read.nddspiro("example.xml", traces=3)

read.nddspiro("example.xml", indices=c("FEV1", "FVC"))

read.nddspiro("example.xml", accepted_only=FALSE, add_GLI=TRUE)

}
\author{
Jamie Rylance, \email{jamie.rylance@lstmed.ac.uk}
}
