% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nddspiro.results.R
\name{nddspiro.results}
\alias{nddspiro.results}
\title{Get spirometry results}
\usage{
nddspiro.results(indices, file, xpath, label = "", test_only = FALSE,
  no_ranges = FALSE)
}
\arguments{
\item{test_only}{Boolean. Optional. If TRUE suppresses the return patient details.}

\item{no_ranges}{Boolean. Optional. If TRUE suppresses the predicted values.}

\item{param}{Character vector. Parameter to be returned.}

\item{nodelist}{XML nodelist. Nodes to be searched.}
}
\value{
Dataframe of results for the given parameters.
}
\description{
Extracts requested measurements from an ndd Easyware XML file
}
\keyword{internal}
