% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nddspiro.values.R
\name{nddspiro.values}
\alias{nddspiro.values}
\title{Get spirometry data from an ndd Easyware XML file}
\usage{
nddspiro.values(param, nodelist, no_ranges = FALSE, label = "")
}
\arguments{
\item{param}{Character vector. Parameter to be returned.}

\item{nodelist}{XML nodelist. Nodes to be searched.}

\item{no_ranges}{Boolean. Optional. If TRUE supresses the return of LLN and predicted values.}

\item{label}{Optional. Character vector. Appended to each of the column names.}
}
\value{
Dataframe of results for the given parameter.
}
\description{
\code{nddspiro.results} returns a dataframe containing the requested measurements.
}
\keyword{internal}
